Restart the service when the configuration file changes too
We set the provider to be debian, since the init.d file is not done via
Upstart - on either Debian or Ubuntu.

Also, set hasstatus to false so that we look through the process list.
The init.d script will return with an exit code of 0, if nothing is
running - which is not what we want.