Default SSL to off.
In Postgres 9.2+, if SSL is configured the server will no longer
silently continue if the server / root certificates and keys are
not where it expects.

Therefore default SSL to off, but in place parameters so that if you
turn it on - and you are using Postgres 9.2+ - you can specify the
particular filenames.