Do not start or manage the Postgresql daemon.
In some situations, for example, a clustered environment when and how to
start Postgresql might be taken by another service or system.

If you specify 'manage_service' to be false, a policy-rc.d(8) shell
script is installed that will prevent the Debian / Ubuntu service from
starting at install time.

Additionally the entire service definition is removed, so that other
things can not trigger the service to run (or stop) unexpectedly.